Improve performance of AddKill function
Previously, every time a call was made to AddKill, it would return the
resulting state of the mob by use of GetKill. GetKill iterates *all*
mobs in KillTrack's database to find the requested mob. This means,
every time a kill is added, the mob database is iterated.

In situations with a large of deaths occurring, like in a raid, this
could lead to degraded performance.

Instead, the code now simply grabs the mob data directly from the
saved variables table with the ID, so no iteration is needed.