Fix bug with game menu
When calling `ShowUIPanel(InterfaceOptionsFrame)` and then closing out
of the interface options, the game menu can no longer be opened by
pressing the ESC key, forcing a reload of the UI to fix it. (Opening it
with the button in the micromenu also works to make it unstuck.)

Change the behaviour of the options helper to instead open to the
KillTrack options with the old workaround of calling
InterfaceOptionsFrame_OpenToCategory twice.