Fix built-in script being overridden
For the search box in the mob list, the old OnTextChanged script handler
caused the Blizzard built-in handler to not run, making the placeholder
text in the search box not hide and overlay the actual query the user
wrote.

This appears to be due to a change in how the SearchBoxTemplate works
as the existing code had worked previously.