Fix tooltip behaviour on classic clients
When the mouse leaves a unit on classic clients,
the UPDATE_MOUSEOVER_UNIT event fires, unlike on retail.

This led to the tooltip text being duplicated whenever the user moved
the mouse off a unit, and the tooltip staying on screen because of the
`Show` method on `GameTooltip` being called after it should be hidden.

Hooking the `OnTooltipSetUnit` script on `GameTooltip` instead proves
to be a more reliable method that works across all clients.